% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UStatsMain.R
\name{KMUTest}
\alias{KMUTest}
\title{A U-statistics method for association test between a set of predictors and multiple outcomes}
\usage{
KMUTest(x, y, z, pred = "Cont")
}
\arguments{
\item{x}{A n * p matrix of predictors with n subjects and p predictors.}

\item{y}{A n * p matrix of outcomes with n subjects and p phenotypes.}

\item{z}{A matrix of demographic variables (e.g., age and gender).}

\item{pred}{Specify the types of predictors ("Cont": Continuous predictors, where linear, quadratic and gaussian kernels are used. "Cat": categorical predictors, where linear, weighted linear with beta weights, and quadratic kernels are used.)}
}
\value{
The P-value of the test.
}
\description{
A U-statistics method for association test between a set of predictors and multiple outcomes
}
\examples{
data(Example)
KMUTest(xcont,y,z,pred="Cont")
KMUTest(xcat,y,z,pred="Cat")
}
